Fix crash caused by race condition on freeing plugin BinaryViews
`FreeObjectReference` on `BinaryView` can be deferred to not block the
UI on teardown of large binary views. This causes
`ReleaseForRegistration` to be called at a later time, which could cause
a race that permitted two threads to delete the object at the same time.
Moving the `FreeObjectReference` to be after the API reference release
causes `ReleaseForRegistration` to be always called after all API
references have been completed, so it will no longer free twice.